// src/ts/types/gprs.types.ts
export type GPRStype = {
  ip: string;
  port: string;
  deviceId: string;
  folderId?: string;
};

export type DinRailType = {
  ip: string;
  port: string;
  deviceId: string;
  password: string;
  type: "ONE_PHASE" | "THREE_PHASE";
};

export type OtanType = {
  ip: string;
  port: string;
  deviceId: string;
  password: string;
};

export type GPRSTableData = {
  dateFrom: DateType;
  dateTo: DateType;
  id: Array<string>;
  type: string;
};
// ...
export type GPRSDescription = {
  ip: string;
  port: number;
  deviceId: string;
  manufacturer: string;
  setUpOrganization: string;
  meterName: string;
  verificationDate: DateType;
  verificationTime: string;
};

export type OtanAttributesType = {
  ip: string;
  port: number;
  deviceId: string;
  accuracyClassActiveEnergy: string;
  accuracyClassReactiveEnergy: string;
};

export type GPRSInfoType = {
  ip: string;
  port: string;
  deviceId: string;
  personalAccountNumber: string;
  roleName: string;
  organizationName: string;
  firstname: string;
  lastname: string;
  fathersname: string;
 
  position: string;
  phoneNumber: string;
  email: string;
  city: string;
  district: string;
  street: string;
  house: string;
  flat: string;
};